Update UI styling and implement arithmetic category in Kafka service
Several key changes have been made. The UI of the Kafka game page in the app was updated with added shadow and border to improve its appearance. In the Kotlin service, scoring for existing categories has been adjusted and an arithmetic category has been introduced, which is handled in the Kafka processor. Thus, this enhances the functionality of the Kafka service by processing arithmetic tasks.